+++ apps/fluent-tester/src/testPages.ts
@@ -6,6 +6,7 @@ import { ButtonTest } from './TestComponents/Button';
 import { CalloutTest } from './TestComponents/Callout';
 import { CheckboxLegacyTest } from './TestComponents/CheckboxLegacy';
 import { CheckboxV1Test } from './TestComponents/CheckboxV1';
+import { ColorTokensTest } from './TestComponents/ColorTokens';
 import { ContextualMenuTest } from './TestComponents/ContextualMenu';
 import { CornerRadiusTokensTest } from './TestComponents/CornerRadius';
 import { NativeDatePickerTest, HOMEPAGE_NATIVEDATEPICKER_BUTTON } from './TestComponents/NativeDatePicker';
@@ -31,14 +32,13 @@ import { ShadowTest } from './TestComponents/Shadow';
 import { ShimmerTest } from './TestComponents/Shimmer';
 import { SpacingTokensTest } from './TestComponents/Spacing';
 import { StrokeWidthTest } from './TestComponents/StrokeWidth';
-import { SvgTest } from './TestComponents/Svg';
+import { SvgTest, RNSVGIconsTest } from './TestComponents/Svg';
 import { SwitchTest } from './TestComponents/Switch';
 import { TabsLegacyTest } from './TestComponents/TabsLegacy';
 import { TabsV1Test } from './TestComponents/TabsV1';
 import { TextLegacyTest } from './TestComponents/TextLegacy';
 import { TextV1Test } from './TestComponents/TextV1';
 import { ThemeTest } from './TestComponents/Theme';
-import { TokenTest } from './TestComponents/Tokens';
 import * as Constants from '../../E2E/src/index.consts';
 
 export const tests: TestDescription[] = [
@@ -91,6 +91,12 @@ export const tests: TestDescription[] = [
     platforms: ['android', 'ios', 'macos', 'win32'], // 'windows': GH#935: Temporarily disabling while SVGs don't work in windows
   },
   {
+    name: 'Color Tokens',
+    component: ColorTokensTest,
+    testPageButton: Constants.HOMEPAGE_COLORTOKEN_BUTTON,
+    platforms: ['android', 'ios', 'macos', 'win32', 'windows'],
+  },
+  {
     name: 'ContextualMenu',
     component: ContextualMenuTest,
     testPageButton: Constants.HOMEPAGE_CONTEXTUALMENU_BUTTON,
@@ -131,7 +137,7 @@ export const tests: TestDescription[] = [
     name: 'Focus Trap Zone',
     component: FocusTrapTest,
     testPageButton: Constants.HOMEPAGE_FOCUSTRAPZONE_BUTTON,
-    platforms: ['android', 'win32'],
+    platforms: ['win32'],
   },
   {
     name: 'FocusZone',
@@ -209,7 +215,7 @@ export const tests: TestDescription[] = [
     name: 'RadioGroup V1',
     component: RadioGroupV1Test,
     testPageButton: Constants.HOMEPAGE_RADIOGROUPV1_BUTTON,
-    platforms: ['macos', 'win32'],
+    platforms: ['android', 'macos', 'win32'],
   },
   {
     name: 'Separator',
@@ -242,9 +248,15 @@ export const tests: TestDescription[] = [
     platforms: ['android', 'ios', 'macos', 'win32', 'windows'],
   },
   {
-    name: 'Svg',
+    name: 'Svg rendering',
     component: SvgTest,
     testPageButton: Constants.HOMEPAGE_SVG_BUTTON,
+    platforms: ['android', 'ios', 'macos', 'win32'],
+  },
+  {
+    name: 'Svg Icon packages',
+    component: RNSVGIconsTest,
+    testPageButton: Constants.HOMEPAGE_RNSVGIcons_BUTTON,
     platforms: ['android', 'ios', 'macos', 'win32'],
   },
   {
@@ -281,12 +293,6 @@ export const tests: TestDescription[] = [
     name: 'Theme',
     component: ThemeTest,
     testPageButton: Constants.HOMEPAGE_THEME_BUTTON,
-    platforms: ['android', 'ios', 'macos', 'win32', 'windows'],
-  },
-  {
-    name: 'Tokens',
-    component: TokenTest,
-    testPageButton: Constants.HOMEPAGE_TOKEN_BUTTON,
     platforms: ['android', 'ios', 'macos', 'win32', 'windows'],
   },
 ];