+++ packages/components/Separator/src/Separator.types.ts
@@ -4,7 +4,7 @@ export const separatorName = 'Separator';
 
 export const SeparatorInsetSpacings = [0, 16, 56, 68, 72, 108] as const;
 
-export type SeparatorInsetSpacing = typeof SeparatorInsetSpacings[number];
+export type SeparatorInsetSpacing = (typeof SeparatorInsetSpacings)[number];
 
 export interface SeparatorPropTokens {
   /**