+++ packages/components/Pressable/CHANGELOG.md
@@ -1,8 +1,48 @@
 # Change Log - @fluentui-react-native/pressable
 
-This log was last generated on Wed, 04 Jan 2023 16:55:53 GMT and should not be manually modified.
+This log was last generated on Fri, 20 Jan 2023 18:06:46 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.9.51
+
+Fri, 20 Jan 2023 18:06:46 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/interactive-hooks to v0.22.12
+
+## 0.9.50
+
+Wed, 18 Jan 2023 01:55:11 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/interactive-hooks to v0.22.11
+
+## 0.9.49
+
+Tue, 10 Jan 2023 20:17:18 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/interactive-hooks to v0.22.10
+
+## 0.9.48
+
+Mon, 09 Jan 2023 22:25:00 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/interactive-hooks to v0.22.9
+
+## 0.9.47
+
+Fri, 06 Jan 2023 12:26:31 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/interactive-hooks to v0.22.8
 
 ## 0.9.46