+++ packages/deprecated/theming-ramp/CHANGELOG.md
@@ -1,8 +1,24 @@
 # Change Log - @uifabricshared/theming-ramp
 
-This log was last generated on Thu, 29 Dec 2022 03:47:47 GMT and should not be manually modified.
+This log was last generated on Wed, 18 Jan 2023 01:55:11 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.17.16
+
+Wed, 18 Jan 2023 01:55:11 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/theme-types to v0.29.1
+
+## 0.17.15
+
+Tue, 10 Jan 2023 20:17:18 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/theme-types to v0.29.0
 
 ## 0.17.14